moving special representer lines to StackDiffer
